add loshdex py package
