+++ py/pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "loshdex"
+version = "0.0.1"
+description = "this is helper for https://github.com/attentionmech/dex project"
+authors = [{ name="attentionmech", email="attentionmech@gmail.com" }]
+license = "Apache 2.0"
+readme = "README.md"
+requires-python = ">=3.8"
+
+dependencies = [
+    "torch==2.5.0",
+    "transformers",
+    "huggingface_hub",
+    "pandas",
+    "pyarrow"
+]
+
+[project.scripts]
+dex-cli = "loshdex.core:main"
+
